CLN: move SHG-PL-Trans package into addons module
