+++ pyfibre/addons/shg_pl_trans/tests/fixtures/test-pyfibre-pl-Stack.tif
