+++ pyfibre/addons/shg_pl_trans/tests/probe_classes.py
@@ -1,9 +1,10 @@
 import numpy as np
 from skimage.io import imread
 
-from pyfibre.shg_pl_trans.shg_image import SHGImage
-from pyfibre.shg_pl_trans.shg_pl_trans_image import SHGPLTransImage
-from pyfibre.shg_pl_trans.tests.fixtures import (
+from pyfibre.addons.shg_pl_trans.shg_image import SHGImage
+from pyfibre.addons.shg_pl_trans.shg_pl_trans_image import (
+    SHGPLTransImage)
+from pyfibre.addons.shg_pl_trans.tests.fixtures import (
     test_shg_pl_trans_image_path,
     test_shg_image_path
 )
