+++ pyfibre/addons/shg_pl_trans/tests/test_shg_pl_trans_analyser.py
@@ -8,14 +8,15 @@ from pyfibre.core.pyfibre_runner import PyFibreRunner
 from pyfibre.model.objects.segments import (
     FibreSegment, CellSegment
 )
-from pyfibre.shg_pl_trans.tests.probe_classes import ProbeSHGPLTransImage
-from pyfibre.shg_pl_trans.shg_pl_trans_analyser import (
+from pyfibre.addons.shg_pl_trans.shg_pl_trans_analyser import (
     SHGPLTransAnalyser)
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
 from pyfibre.tests.probe_classes.objects import (
     ProbeFibreNetwork, generate_probe_segment)
 from pyfibre.tests.probe_classes.utilities import (
     generate_image, generate_probe_graph)
+
+from .probe_classes import ProbeSHGPLTransImage
 
 
 class TestSHGPLTransAnalyser(PyFibreTestCase):
