+++ pyfibre/addons/shg_pl_trans/tools/segmentation.py
