+++ pyfibre/addons/shg_pl_trans/tools/tests/test_segmentation.py
@@ -4,8 +4,9 @@ from pyfibre.tests.probe_classes.utilities import (
     generate_image, generate_probe_graph)
 from pyfibre.tests.probe_classes.objects import ProbeFibreNetwork
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
-from pyfibre.shg_pl_trans.tests.probe_classes import ProbeSHGPLTransImage
-from pyfibre.shg_pl_trans.tests.fixtures import (
+from pyfibre.addons.shg_pl_trans.tests.probe_classes import (
+    ProbeSHGPLTransImage)
+from pyfibre.addons.shg_pl_trans.tests.fixtures import (
     test_shg_pl_trans_image_path)
 
 from ..segmentation import (
