+++ pyfibre/addons/shg_pl_trans/tools/tests/test_shg_pl_trans_kmeans_filter.py
@@ -1,13 +1,13 @@
 import numpy as np
 from skimage.io import imread
 
-from pyfibre.shg_pl_trans.tools.shg_pl_trans_kmeans_filter import (
+from pyfibre.addons.shg_pl_trans.tools.shg_pl_trans_kmeans_filter import (
     nonzero_mean, spherical_coords,
     binary_classifier_spherical,
     distance_sum,
     SHGPLTransKmeansFilter
 )
-from pyfibre.shg_pl_trans.tests.fixtures import (
+from pyfibre.addons.shg_pl_trans.tests.fixtures import (
     test_shg_pl_trans_image_path)
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
